chore: cleanup API
